+++ packages/@tinacms/graphql/src/database/index.ts
@@ -531,7 +531,13 @@ export class Database {
 
         // If a collection match is specified, make sure the file matches the glob.
         // TODO: Maybe we should service this error better in the frontend?
-        if (collection.match?.exclude || collection.match?.include) {
+        const isFolderPlaceholder = filepath.endsWith(
+          `.gitkeep.${collection.format || 'md'}`
+        );
+        if (
+          !isFolderPlaceholder &&
+          (collection.match?.exclude || collection.match?.include)
+        ) {
           const matches = this.tinaSchema.getMatches({ collection });
 
           const match = micromatch.isMatch(filepath, matches);