+++ packages/tinacms/src/toolkit/fields/components/color-picker/color-picker.tsx
@@ -12,7 +12,7 @@ interface FieldInput {
   onChange: (value: string | null) => void;
 }
 
-interface SwatchProps extends React.HTMLAttributes<HTMLDivElement> {
+interface SwatchProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {
   colorRGBA?: ColorRGBA;
   onClick: (_event: React.SyntheticEvent) => void;
   colorFormat: ColorFormat;
@@ -34,9 +34,10 @@ const getTextColorForBackground = function (backgroundColor?: ColorRGBA) {
 };
 
 const Swatch = ({ colorRGBA, colorFormat, width, ...props }: SwatchProps) => (
-  <div
-    className='bg-gray-100 rounded shadow-[0_2px_3px_rgba(0,0,0,0.12)] cursor-pointer m-0'
-    style={{ width }}
+  <button
+    type='button'
+    className='block bg-gray-100 rounded shadow-[0_2px_3px_rgba(0,0,0,0.12)] m-0'
+    style={{ width: width || '100%' }}
     {...props}
   >
     <div
@@ -53,7 +54,7 @@ const Swatch = ({ colorRGBA, colorFormat, width, ...props }: SwatchProps) => (
         ? 'Click to add color'
         : ColorFormatter[colorFormat].getLabel(colorRGBA)}
     </div>
-  </div>
+  </button>
 );
 
 const Popover = ({